+++ eras/dijkstra/impl/cddl/lib/Cardano/Ledger/Dijkstra/HuddleSpec.hs
@@ -447,6 +447,7 @@ dijkstraCertificateRule ::
   ( HuddleGroup "delegation_to_stake_pool_cert" era
   , HuddleGroup "pool_registration_cert" era
   , HuddleGroup "pool_retirement_cert" era
+  , HuddleGroup "bls_key_registration_cert" era
   , HuddleGroup "account_registration_deposit_cert" era
   , HuddleGroup "account_unregistration_deposit_cert" era
   , HuddleGroup "delegation_to_drep_cert" era
@@ -480,6 +481,7 @@ dijkstraCertificateRule pname p =
     / arr [a $ huddleGroup @"drep_registration_cert" p]
     / arr [a $ huddleGroup @"drep_unregistration_cert" p]
     / arr [a $ huddleGroup @"drep_update_cert" p]
+    / arr [a $ huddleGroup @"bls_key_registration_cert" p]
 
 instance HuddleRule "bounded_bytes" DijkstraEra where
   huddleRuleNamed pname _ = boundedBytesRule pname
@@ -625,13 +627,24 @@ instance HuddleRule "bls_key" DijkstraEra where
               , TBytes (rawEncodeFixedSized $ blsPossessionProof lk)
               ]
 
+instance HuddleGroup "bls_key_registration_cert" DijkstraEra where
+  huddleGroupNamed pname p =
+    ( pname
+        =.~ grp
+          [ 19
+          , a $ huddleRule @"pool_keyhash" p
+          , "bls_pubkey" ==> VBytes `sized` (96 :: Word64)
+          , "bls_possession_proof" ==> VBytes `sized` (48 :: Word64)
+          ]
+    )
+      //- "Register or rotate a stake pool's Leios voting key"
+
 instance HuddleGroup "pool_params" DijkstraEra where
   huddleGroupNamed pname p =
     ( pname
         =.~ grp
           [ "operator" ==> huddleRule @"pool_keyhash" p
           , "vrf_keyhash" ==> huddleRule @"vrf_keyhash" p
-          , opt ("bls_key" ==> huddleRule @"bls_key" p / VNil)
           , "pledge" ==> huddleRule @"coin" p
           , "cost" ==> huddleRule @"coin" p
           , "margin" ==> huddleRule @"unit_interval" p