+++ eras/shelley/test-suite/bench/Cardano/Ledger/Shelley/Bench/Rewards.hs
@@ -154,7 +154,6 @@ genChainInEpoch epoch = do
                         , sppOwners = Set.singleton owner
                         , sppRelays = StrictSeq.empty
                         , sppMetadata = SNothing
-                        , sppBlsKey = SNothing
                         }
               ]
         , sgsStake =